Небольшие переделки по модульности
